Fix: add to cart (#110)
* Update cart_api.ts

* Fix: Add user ID to cart item creation

The frontend only sent doll_id and quantity when adding items to the cart, resulting in a 422 error. This commit adds the user ID to the request to fix the issue.

* Fix: Fetch doll details in cart and checkout

Fetch doll details using the doll ID from the cart API to display model information in the cart and checkout pages.

* Delete .env

* Update index.html

* Delete package-lock.json

* Apply suggestions from code review

--------

Co-authored-by: Elvis Otieno <125451537+the1Riddle@users.noreply.github.com>