this is what i want (#162)
* feat: Fetch user role and membership

Updated auth API to retrieve user role and membership level from backend.

* feat: Display user role and membership level

Updated the user dashboard to display the user's role and membership level fetched from the backend.

* feat: Display membership level in footer and navbar

Updated the Footer and Navbar components to display the user's membership level fetched from the backend API.

* Refactor: Extract StaffNavbar component

Extracts the Navbar component into a StaffNavbar component, and uses it in the Moderator and Admin pages.

---------

Co-authored-by: the-1Riddle <154701770+the-1Riddle@users.noreply.github.com>